refactor(bitwise.ts): merge rotate function from rot.ts into bitwise.ts
